integrated price info with product details

+++ src/main/java/com/test/trgt/api/GatewayApi.java
@@ -7,6 +7,8 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import com.test.trgt.models.Error;
+import com.test.trgt.models.Price;
+import com.test.trgt.data.PricesRepository;
 import com.test.trgt.exceptions.BaseException;
 import com.test.trgt.models.Product;
 import com.test.trgt.service.ProductService;
@@ -24,14 +26,15 @@ public class GatewayApi {
 	@GET
 	@Path("/")
 	public Response resp(){
+		
 		return Response.status(200).entity("HOME PATH !").build();
 	}
 
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/products/{id}")
-	public Product responseMsg(@PathParam("id") long productId) {
-
+	public Product responseMsg(@PathParam("id") int productId) {
+		
 		Product product = new Product();
 		try {
 			product = productService.getProductDetailsById(productId);
